Move Django helpers to DjangoSearch
