+++ code_annotations/find_django.py
@@ -5,11 +5,14 @@ import inspect
 import os
 import pprint
 import re
+import sys
 
+import django
+from django.apps import apps
+from django.db import models
 from six import text_type
 
 from code_annotations.base import BaseSearch
-from code_annotations.django_reporting_helpers import get_model_id, get_models_requiring_annotations
 from code_annotations.helpers import fail, get_annotation_regex, yaml_ordered_dump, yaml_ordered_load
 
 DEFAULT_SAFELIST_FILE_PATH = '.annotation_safe_list.yml'
@@ -28,7 +31,7 @@ class DjangoSearch(BaseSearch):
             config: Configuration file path
         """
         super(DjangoSearch, self).__init__(config)
-        self.local_models, self.non_local_models = get_models_requiring_annotations()
+        self.local_models, self.non_local_models = self.get_models_requiring_annotations()
 
     def seed_safelist(self):
         """
@@ -42,7 +45,7 @@ class DjangoSearch(BaseSearch):
                 len(self.non_local_models))
         )
 
-        safelist_data = {get_model_id(model): {} for model in self.non_local_models}
+        safelist_data = {self.get_model_id(model): {} for model in self.non_local_models}
 
         with open(self.config.safelist_path, 'w') as safelist_file:
             yaml_ordered_dump(safelist_data, stream=safelist_file)
@@ -60,7 +63,7 @@ class DjangoSearch(BaseSearch):
             self.echo(
                 'Listing {} local models requiring annotations:'.format(len(self.local_models))
             )
-            pprint.pprint(sorted([get_model_id(model) for model in self.local_models]), indent=4)
+            pprint.pprint(sorted([self.get_model_id(model) for model in self.local_models]), indent=4)
         else:
             self.echo('No local models requiring annotations.')
 
@@ -88,7 +91,7 @@ class DjangoSearch(BaseSearch):
         # TODO: Untangle this nest
         # pylint: disable=too-many-nested-blocks
         for model in self.local_models.union(self.non_local_models):
-            model_id = get_model_id(model)
+            model_id = self.get_model_id(model)
             hierarchy = inspect.getmro(model)
             model_annotations = []
 
@@ -109,7 +112,7 @@ class DjangoSearch(BaseSearch):
 
                             if len(cleaned_groups) != 2:  # pragma: no cover
                                 raise Exception('{}: Number of found items in the list is not 2. Found: {}'.format(
-                                    get_model_id(obj),
+                                    self.get_model_id(obj),
                                     cleaned_groups
                                 ))
 
@@ -157,3 +160,107 @@ class DjangoSearch(BaseSearch):
                     })
                 self.format_file_results(annotated_models, [model_annotations])
         return annotated_models
+
+    @staticmethod
+    def requires_annotations(model):
+        """
+        Return true if the given model actually requires annotations, according to PLAT-2344.
+        """
+        # Anything inheriting from django.models.Model will have a ._meta attribute. Our tests
+        # inherit from object, which doesn't have it, and will fail below. This is a quick way
+        # to early out on both.
+        if not hasattr(model, '_meta'):
+            return False
+
+        return issubclass(model, models.Model) \
+            and not (model is models.Model) \
+            and not model._meta.abstract \
+            and not model._meta.proxy
+
+    @staticmethod
+    def is_non_local(model):
+        """
+        Determine if the given model is non-local to the current IDA.
+
+        Non-local models are all installed models which are not "local", by
+        definition.  "Local" models are installed models which are defined in code
+        which physically lives in the current codebase, where "current codebase"
+        ostensibly refers to the code providing the currently active django
+        project.
+
+        Args:
+            model (django.db.models.Model): A model to check.
+
+        Returns:
+            bool: True if the given model is non-local.
+        """
+        # If the model _was_ local to the current IDA repository, it should be
+        # defined somewhere under sys.prefix + '/src/' or in a path that points to
+        # the current checked-out code.  On Posix systems according to our testing,
+        # non-local packages get installed to paths containing either
+        # "site-packages" or "dist-packages".
+        non_local_path_prefixes = []
+        for path in sys.path:
+            if 'dist-packages' in path or 'site-packages' in path:
+                non_local_path_prefixes.append(path)
+        model_source_path = inspect.getsourcefile(model)
+        return model_source_path.startswith(tuple(non_local_path_prefixes))
+
+    @staticmethod
+    def get_model_id(model):
+        """
+        Construct the django standard model identifier in "app_label.ModelClassName" notation.
+
+        Args:
+            model (django.db.models.Model): A model for which to create an identifier.
+
+        Returns:
+            str: identifier string for the given model.
+        """
+        return '{}.{}'.format(model._meta.app_label, model._meta.object_name)
+
+    @staticmethod
+    def setup_django():
+        """
+        Prepare to make django library function calls.
+
+        On behalf of the current Django project in the current working directory,
+        setup/load the django framework, specified settings, and apps therein.
+        This should be called before calling any django submodule functions which
+        expect apps to be loaded.
+
+        This function is idempotent.
+        """
+        if sys.path[0] != '':  # pragma: no cover
+            sys.path.insert(0, '')
+        django.setup()
+
+    @staticmethod
+    def get_models_requiring_annotations():
+        """
+        Determine all local and non-local models via django model introspection.
+
+        Note that non-local models returned may contain 1st party models (authored by
+        edX).  This is a compromise in accuracy in order to simplify the generation
+        of this list, and also to ease the transition from zero to 100% annotations
+        in edX satellite repositories.
+
+        Returns:
+            tuple:
+                2-tuple where the first item is a set of local models, and the
+                second item is a set of non-local models.
+        """
+        DjangoSearch.setup_django()
+        local_models = set()
+        non_local_models = set()
+        for app in apps.get_app_configs():
+            for root_model in app.get_models():
+                # getmro() includes the _entire_ inheritance closure, not just the direct inherited classes.
+                heirarchy = inspect.getmro(root_model)
+                for model in heirarchy:
+                    if DjangoSearch.requires_annotations(model):
+                        if DjangoSearch.is_non_local(model):
+                            non_local_models.add(model)
+                        else:
+                            local_models.add(model)
+        return local_models, non_local_models
