+++ tests/test_find_django.py
@@ -3,11 +3,17 @@
 """
 Tests for find annotations in Django models.
 """
+import sys
+
 from mock import DEFAULT, patch
 
-from code_annotations.django_reporting_helpers import get_model_id
+from code_annotations.find_django import DjangoSearch
 from tests.fake_models import (
+    FakeBaseModelAbstract,
+    FakeBaseModelBoring,
+    FakeBaseModelBoringWithAnnotations,
     FakeBaseModelNoAnnotation,
+    FakeBaseModelProxy,
     FakeBaseModelWithAnnotation,
     FakeBaseModelWithNoDocstring,
     FakeChildModelMultiAnnotation,
@@ -19,7 +25,7 @@ from tests.helpers import EXIT_CODE_FAILURE, EXIT_CODE_SUCCESS, call_script_isol
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_simple_success(**kwargs):
@@ -44,7 +50,7 @@ def test_find_django_simple_success(**kwargs):
             Raw text contents of the generated report file
         """
         for model in test_models:
-            assert 'object_id: {}'.format(get_model_id(model)) in report_contents
+            assert 'object_id: {}'.format(DjangoSearch.get_model_id(model)) in report_contents
 
     result = call_script_isolated(
         ['django_find_annotations', '--config_file', 'test_config.yml', '--lint', '--report'],
@@ -58,7 +64,7 @@ def test_find_django_simple_success(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_no_viable_models(**kwargs):
@@ -82,7 +88,7 @@ def test_find_django_no_viable_models(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_model_not_annotated(**kwargs):
@@ -107,7 +113,7 @@ def test_find_django_model_not_annotated(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_model_in_safelist_not_annotated(**kwargs):
@@ -139,7 +145,7 @@ def test_find_django_model_in_safelist_not_annotated(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_model_in_safelist_annotated(**kwargs):
@@ -171,7 +177,7 @@ def test_find_django_model_in_safelist_annotated(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_no_safelist(**kwargs):
@@ -192,7 +198,7 @@ def test_find_django_no_safelist(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_in_safelist_and_annotated(**kwargs):
@@ -208,7 +214,7 @@ def test_find_django_in_safelist_and_annotated(**kwargs):
 
     result = call_script_isolated(
         ['django_find_annotations', '--config_file', 'test_config.yml', '--lint', '--report'],
-        fake_safelist_data='{{{}: ".. no_pii::"}}'.format(get_model_id(FakeBaseModelWithAnnotation))
+        fake_safelist_data='{{{}: ".. no_pii::"}}'.format(DjangoSearch.get_model_id(FakeBaseModelWithAnnotation))
     )
 
     assert result.exit_code == EXIT_CODE_FAILURE
@@ -218,7 +224,7 @@ def test_find_django_in_safelist_and_annotated(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_no_docstring(**kwargs):
@@ -242,7 +248,7 @@ def test_find_django_no_docstring(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_ordering_error(**kwargs):
@@ -265,7 +271,7 @@ def test_find_django_ordering_error(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_without_linting(**kwargs):
@@ -289,7 +295,7 @@ def test_find_django_without_linting(**kwargs):
 
 
 @patch.multiple(
-    'code_annotations.find_django',
+    'code_annotations.find_django.DjangoSearch',
     get_models_requiring_annotations=DEFAULT
 )
 def test_find_django_without_report(**kwargs):
@@ -310,3 +316,106 @@ def test_find_django_without_report(**kwargs):
     assert result.exit_code == EXIT_CODE_SUCCESS
     assert 'Linting passed without errors' in result.output
     assert 'Generating report to' not in result.output
+
+
+@patch('code_annotations.find_django.issubclass')
+def test_requires_annotations_abstract(mock_issubclass):
+    """
+    Abstract classes should not require annotations
+    """
+    mock_issubclass.return_value = True
+    assert DjangoSearch.requires_annotations(FakeBaseModelAbstract) is False
+
+
+@patch('code_annotations.find_django.issubclass')
+def test_requires_annotations_proxy(mock_issubclass):
+    """
+    Proxy classes should not require annotations
+    """
+    mock_issubclass.return_value = True
+    assert DjangoSearch.requires_annotations(FakeBaseModelProxy) is False
+
+
+@patch('code_annotations.find_django.issubclass')
+def test_requires_annotations_normal(mock_issubclass):
+    """
+    Non-abstract, non-proxy models should require annotations
+    """
+    mock_issubclass.return_value = True
+    assert DjangoSearch.requires_annotations(FakeBaseModelBoring) is True
+
+
+def test_requires_annotations_not_a_model():
+    """
+    Things which are not models should not require annotations
+    """
+    assert DjangoSearch.requires_annotations(dict) is False
+
+
+def test_is_non_local_simple():
+    """
+    Our model is local, should show up as such
+    """
+    assert DjangoSearch.is_non_local(FakeBaseModelAbstract) is False
+
+
+@patch('code_annotations.find_django.inspect.getsourcefile')
+def test_is_non_local_site(mock_getsourcefile):
+    """
+    Try to test the various non-local paths, if the environment allows.
+    """
+
+    # This code is duplicated from the method itself.
+    non_local_path_prefixes = []
+    for path in sys.path:
+        if 'dist-packages' in path or 'site-packages' in path:
+            non_local_path_prefixes.append(path)
+
+    if non_local_path_prefixes:
+        for prefix in non_local_path_prefixes:
+            mock_getsourcefile.return_value = '{}/bar.py'.format(prefix)
+            assert DjangoSearch.is_non_local(FakeBaseModelAbstract) is True
+    else:
+        # If there are no prefixes in the test environment, there's really nothing to do here.
+        pass
+
+
+@patch('code_annotations.find_django.issubclass')
+@patch('code_annotations.find_django.DjangoSearch.setup_django')
+@patch('code_annotations.find_django.DjangoSearch.is_non_local')
+@patch('code_annotations.find_django.django.apps.apps.get_app_configs')
+def test_get_models_requiring_annotations(mock_get_app_configs, mock_is_non_local, mock_setup_django, mock_issubclass):
+    # Lots of fakery going on here. This class mocks Django AppConfigs to deliver our fake models.
+    class FakeAppConfig(object):
+        def get_models(self):
+            return [FakeBaseModelBoring, FakeBaseModelBoringWithAnnotations]
+
+    # This lets us deterministically decide that one model is local, and the other isn't, for testing both branches.
+    mock_is_non_local.side_effect = [True, False]
+
+    # This just fakes setting up Django
+    mock_setup_django.return_value = True
+
+    # This mocks out Django's get_app_configs to return our fake AppConfig
+    mock_get_app_configs.return_value = [FakeAppConfig()]
+
+    # This lets us pretend that all of our fake models inherit from Django's model.Model.
+    # If we try to do that inheritance Django will throw errors unless we do a full Django
+    # testing setup.
+    mock_issubclass.return_value = True
+
+    local_models, non_local_models = DjangoSearch.get_models_requiring_annotations()
+
+    assert len(local_models) == 1
+    assert len(non_local_models) == 1
+    assert list(local_models)[0] == FakeBaseModelBoringWithAnnotations
+    assert list(non_local_models)[0] == FakeBaseModelBoring
+
+
+@patch('code_annotations.find_django.django.setup')
+def test_setup_django(mock_django_setup):
+    """
+    This is really just for coverage.
+    """
+    mock_django_setup.return_value = True
+    DjangoSearch.setup_django()
